Resolves #1 by removing temporary nocite.
For some reason we added nocite and forgot to remove it from the final
submission. It has been removed now. However, I don't fully agree with
the other comment regarding the order of the references. In LaTeX it's
common for the bibliography to be ordered by the first author's names.