Resolves #2, blatant abuse of italics.
I've chosen to only use italics when referring to authors, names of
companies or abbreviations introduced for later chapters. The other
use is when introducing key technical terms (which are non-standad)
like textured appearance or central terms like mesh simplification.
There are now only use once they are introduced as well, not again.